% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{claude_log}
\alias{claude_log}
\title{Simple logging utility for Claude addins}
\usage{
claude_log(message, level = "INFO", addin = "Claude")
}
\arguments{
\item{message}{The message to log}

\item{level}{The log level (INFO, DEBUG, WARNING, ERROR)}

\item{addin}{The name of the addin generating the log}
}
\description{
Simple logging utility for Claude addins
}
