% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{select_model}
\alias{select_model}
\title{Select a Claude Model}
\usage{
select_model(client, model_number)
}
\arguments{
\item{client}{An anthropic_client object}

\item{model_number}{The number ID of the model to use}
}
\value{
The client with the selected model
}
\description{
Select a Claude Model
}
